Rename old arxiv files
