+++ example.cpp
@@ -30,9 +30,9 @@
 #include "get_fitzgibbon_cvpr_2001.hpp"
 #include "get_kukelova_cvpr_2015.hpp"
 #include "get_nakano_icpr_2025.hpp"
-#include "get_valtonenornhag_arxiv_2020a.hpp"
-#include "get_valtonenornhag_arxiv_2020b.hpp"
-#include "get_wadenback_2025.hpp"
+// #include "get_valtonenornhag_icpr_2020.hpp"
+// #include "get_valtonenornhag_wacv_2021.hpp"
+#include "get_wadenback_3dv_2026.hpp"
 #include "problem_instance.hpp"
 #include "generate_problem_instance.hpp"
 #include "posedata.hpp"
