+++ src/solvers/valtonenornhag_wacv_2021/solver_valtonenornhag_wacv_2021_fHf.hpp
@@ -18,15 +18,15 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#ifndef SRC_SOLVERS_VALTONENORNHAG_ARXIV_2020A_SOLVER_VALTONENORNHAG_ARXIV_2020A_FHF_HPP_
-#define SRC_SOLVERS_VALTONENORNHAG_ARXIV_2020A_SOLVER_VALTONENORNHAG_ARXIV_2020A_FHF_HPP_
+#ifndef SRC_SOLVERS_VALTONENORNHAG_WACV_2021_SOLVER_VALTONENORNHAG_WACV_2021_FHF_HPP_
+#define SRC_SOLVERS_VALTONENORNHAG_WACV_2021_SOLVER_VALTONENORNHAG_WACV_2021_FHF_HPP_
 
 #include <Eigen/Dense>
 
 namespace HomLib {
-namespace ValtonenOrnhagArxiv2020A {
-Eigen::MatrixXcd solver_fHf(const Eigen::VectorXd& data);
+namespace ValtonenOrnhagWACV2021 {
+Eigen::VectorXcd solver_fHf(const Eigen::VectorXd& data);
 }
 }
 
-#endif  // SRC_SOLVERS_VALTONENORNHAG_ARXIV_2020A_SOLVER_VALTONENORNHAG_ARXIV_2020A_FHF_HPP_
+#endif  //  SRC_SOLVERS_VALTONENORNHAG_WACV_2021_SOLVER_VALTONENORNHAG_WACV_2021_FHF_HPP_
