+++ src/solvers/valtonenornhag_wacv_2021/solver_valtonenornhag_wacv_2021_frHfr.hpp
@@ -18,13 +18,14 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#ifndef SRC_SOLVERS_VALTONENORNHAG_ARXIV_2020B_SOLVER_VALTONENORNHAG_ARXIV_2020B_FRHFR_HPP_
-#define SRC_SOLVERS_VALTONENORNHAG_ARXIV_2020B_SOLVER_VALTONENORNHAG_ARXIV_2020B_FRHFR_HPP_
+#ifndef SRC_SOLVERS_VALTONENORNHAG_WACV_2021_SOLVER_VALTONENORNHAG_WACV_2021_FRHFR_HPP_
+#define SRC_SOLVERS_VALTONENORNHAG_WACV_2021_SOLVER_VALTONENORNHAG_WACV_2021_FRHFR_HPP_
+
 #include <Eigen/Dense>
 
 namespace HomLib {
-namespace ValtonenOrnhagArxiv2020B {
+namespace ValtonenOrnhagWACV2021 {
 Eigen::MatrixXcd solver_frHfr(const Eigen::VectorXd& data);
 }
 }
-#endif  // SRC_SOLVERS_VALTONENORNHAG_ARXIV_2020B_SOLVER_VALTONENORNHAG_ARXIV_2020B_FRHFR_HPP_
+#endif  // SRC_SOLVERS_VALTONENORNHAG_WACV_2021_SOLVER_VALTONENORNHAG_WACV_2021_FRHFR_HPP_
